Rescale Pixel Factor to to something more manageable
The old scale between 30 and 10000 was not user friendly and it was inverted (bigger value got smaller stroke)

Now the default scale is 1.0 and bigger values get thickerr strokes.

+++ source/blender/draw/engines/gpencil/gpencil_engine.h
@@ -161,7 +161,7 @@ typedef struct GPENCIL_Storage {
 	int xray;
 	int keep_size;
 	float obj_scale;
-	int pixfactor;
+	float pixfactor;
 	int playing;
 	bool is_render;
 	int uselines;
